renamed fixture.py to conftest.py and modified the admin_user fixture
