+++ theinkspot/users/tests/conftest.py
@@ -1,5 +1,17 @@
 import pytest
 
+from theinkspot.users.models import User
+
+"""
+This module will provide fixtures for the entire directory,
+Fixtures defined here will be used in any test automatically
+without importing them *according to pytest-django documentation*
+"""
+
+# modified the admin user fixture to add the "name" field to the user_data
+# before passing it the create_superuser function because it is one of the
+# required fields in the user model
+
 
 @pytest.fixture()
 def admin_user(
@@ -25,8 +37,29 @@ def admin_user(
         user_data = {}
         if "email" in UserModel.REQUIRED_FIELDS:
             user_data["email"] = "admin@example.com"
+        if "name" in UserModel.REQUIRED_FIELDS:
+            user_data["name"] = "user full name"
         user_data["password"] = "password"
-        user_data["name"] = "admin_name"  # TODO this is the fix
         user_data[username_field] = username
         user = UserModel._default_manager.create_superuser(**user_data)
     return user
+
+
+@pytest.fixture
+def user(db) -> User:
+    return User.objects.create_user(
+        name="user name",
+        username="username",
+        email="test@email.com",
+        password="Am0123456789123456",
+    )
+
+
+@pytest.fixture
+def superuser(db) -> User:
+    return User.objects.create_superuser(
+        name="user name",
+        username="username",
+        email="test@email.com",
+        password="Am0123456789123456",
+    )
